+++ RELEASE_NOTES.md
@@ -59,6 +59,8 @@
 
 - Updated the package name of `location.proto` to `frequenz.api.common.v1`.
 
+- Added messages to support pagination in APIs.
+
 ## New Features
 
 <!-- Here goes the main new features and examples or instructions on how to use them -->